New file structure for Carmin Server (#6)
* New file structure for Carmin Server

This new file structure is based on Flask good practices mixed with the good elements of the Swagger code-gen from the OpenAPI 2.0 version.

* Loosen travis linter policy

We will keep strict linters in our code editor, but loosen the policy on travis. This is done since the linter suggests things that are contrary to best practices in Flask and pytest, especially as it relates to decorators.