SQLAlchemy and Marshmallow (#11)
* Begin adding models

* Start authentication

* Adding PlatformProperties model

* Add remaining models

* Further model code-gen adaptation

* Convert docs and add default_value attribute

* Classes first review from code-gen version

Adaptation for Open Api 3.0

* Moving carmin-server out of cbrain

Marked six as a dependency as it is used in base_model for to_dict.

* Multiple fixes and continued models

Fixed type in documentation (string to str)
Removed cbrain_server_url
Fixed path_md5, path and upload_data from to_be_verified

SIDENOTE:
upload_data had a member called type, which overwrites the built-in type. I renamed it to upload_type within the class to avoid this. However, the attribute_map still refers to type.

* Restructure of the app according to Flask-RESTful (#10)

* Restructure of the app according to Flask-RESTful

Reworking of the app to use the Flask-RESTful library.

* Fixing travis paths

* Add resources and endpoints

* All endpoints and resources complete

* Fix imports to account for new dir structure

* Models continued (DirectoryList)

I put GetPathResponse on stand by for now as it is only required in one endpoint and could not fint an resources as to how to represent it.

* First full request with Flask-SQLAlchemy

First implementation of an endpoitn using SQLAchemy with a SQLite DB. This is just the very first version. A lot of understanding of the library had to be done. The proper setup and structure of the app with the DB has to be reconsidered, but we now have a fonctionnal model and we can return a JSON version of the DB query result (thanks to @marshal_with from flask_restful)

* Remove attribute for as same name

* First full implementation of SQL-Alchemy with JSON loaded and returned properly

* First implementation with marshmallow and custom marshal_with before and after

* Implementing new project structure and moving custom decorators to their own package

* Final demonstration with dump_to and load_from

dump_to and load_from allow us to spedicy the expected variables names from the JSON request to map them properly to the corresponding property

* Moving database related files to /database and /database/models

* Changing config path accordingly