Get stdout stderr (#17)
* Execution Play first changes

Include ExecutionProcess DB models and misc. fixes

* Execution Play part 1 and ErrorCodeAndMessage small rework

* Fix test_put_with_invalid_upload_type

* Properly placing the folder deletion

* Fixing input parameter validation

* Execution play continued

* Further progress in Execution Play

We now load all Boutiques Pipeline and transform them to valid Carmin pipeline at server launch.
TODO: Endpoint /pipelines/reload accessible to admin only which would reload the boutiques pipelines

* dockerignore rename

* Fixing functions, dockerignore and path imports

* Fixing execution

We now have the proper paths for th descriptor and the inputs. We can launch an execution on the current thread. but for some reason the mount does not appear to work. In fact, I try mounting the execution folder on the execution root, but this does not appear to work. More investigation will have to be done regarding this.

* Fixing error message

* Execution taking place

Still need to fix a few things:
- Due to cwd, the mount gives unwanted results when using ../..

* Pipeline and execution fixes

* WIP First admin executions folder

* Load from id for schema

* New admin creation and begin Database refactor

* Change error_detial to accept any type

* Error details fix

* Readme fix- Invalid JSON

* Fixing 4XX that are not user errors

* Full async execution with execution status to DB

* Only initializing executions can be played

* Update command

* Change output file name from processed to greeting

* Change file name

* Boutiques descriptor get first pass (missing tests)

* Tests written. Still 2 tests to fix.

* Get Boutiques descriptor and test fixes done

* Stdout and stderr first version (WIP)

* First functionnal version of stdout and stderr

There is too much code repetition in my opinion. It could be refactored.

* stdout test wip

* Test progress but still experiencing issues with tests setup

* Removing unused import

* Removing old code

* Putting tests on standby

* Commenting out test properly

* Refactor stdout and stderr endpoints

* Change mimetype to text/plain